pyproject.toml
[tool.poetry]
name = "moniker"
version = "1.0.0"
description = "Monitor usernames across various services and report availability via Discord."
authors = ["EthanC <16727756+EthanC@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
loguru-discord = "^1.1.0"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.2"
discord-webhook = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.7"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
